fix(inventory): decimal stock quantities + ledger-true deductions
InventoryItem.Quantity/ReorderLevel/MinimumStock were int while recipes,
transfers and purchases move decimal amounts: every deduction was rounded
(0.2 kg/serving deducted 0; 0.5 deducted 1) and raw SQL moves silently
truncated, so physical stock drifted from the InventoryTransaction ledger.

- Columns widened to decimal(10,2) (matches the HasPrecision config that
  already declared the intent); migration also repairs two check
  constraints missing from live DBs migrated in the reconciliation era.
- InventoryService now moves exact decimal quantities end-to-end; the
  dead UpdateInventoryStatus no-op hook and unused GetInventoryStatus
  helper are gone.
- Low-stock alerts now fire on the automated paths (sales deductions,
  stock-out) when an item CROSSES its reorder/minimum threshold - one
  actionable ping for purchasing instead of silence or spam.
- New IInventoryService.RestockOrderAsync reverses an order''s ledgered
  consumption (idempotent) - groundwork for cancel/compensation flows.
- Deleting an inventory item with movement history/purchases/recipes is
  now blocked instead of orphaning the audit trail.

+++ Models/InventoryItem.cs
@@ -15,9 +15,12 @@ namespace Cafe.Models
         [StringLength(100)]
         public string Name { get; set; } = null!;
 
+        // Stock is measured in the item's Unit (kg, L, pcs, …) and may be fractional —
+        // recipes consume fractional amounts per serving, so an integer column would
+        // silently round every deduction (0.2 kg → 0) and drift from the transaction ledger.
         [Required]
-        [Range(0, int.MaxValue, ErrorMessage = "Quantity cannot be negative.")]
-        public int Quantity { get; set; }
+        [Range(0, double.MaxValue, ErrorMessage = "Quantity cannot be negative.")]
+        public decimal Quantity { get; set; }
 
         [Required]
         [StringLength(20)]
@@ -27,11 +30,11 @@ namespace Cafe.Models
         public int BranchId { get; set; }
 
         [Required]
-        [Range(0, int.MaxValue, ErrorMessage = "Reorder level cannot be negative.")]
-        public int ReorderLevel { get; set; }
+        [Range(0, double.MaxValue, ErrorMessage = "Reorder level cannot be negative.")]
+        public decimal ReorderLevel { get; set; }
 
-        [Range(0, int.MaxValue, ErrorMessage = "Minimum stock cannot be negative.")]
-        public int MinimumStock { get; set; } = 0;
+        [Range(0, double.MaxValue, ErrorMessage = "Minimum stock cannot be negative.")]
+        public decimal MinimumStock { get; set; } = 0;
 
         [StringLength(100)]
         public string? StorageLocation { get; set; }
